Move to using yaml file for configuration
